My_Portfolio-master/requirements.txt
Flask==2.3.2
python-dotenv==1.0.1
gunicorn==21.2.0